backend/pyproject.toml
# [Task]: T-008 [From]: plan.md §Technical Context
# Python project configuration with Ruff for code formatting

[project]
name = "todo-backend"
version = "0.1.0"
description = "FastAPI backend for Todo App Phase II"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.14",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "PyJWT[crypto]>=2.8.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "PyNaCl>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pytest>=7.4.0",
    "httpx>=0.26.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.black]
line-length = 88
target-version = ["py313"]
